+++ Scenes/Scripts/DecorationData.gd
@@ -1,0 +1,26 @@
+var decoration_list = [
+    {
+        "name" : "Locker",
+        "description" :  "+1 to worker cap for all jobs",
+        "show" : [],
+        "building effects" : [["job cap", 1]],
+        "cost" : [],
+        "model" : null
+    },
+    {
+        "name" : "Bin",
+        "description" : "+20% to all resources",
+        "show" : [],
+        "building effects" : [["resource", .20]],
+        "cost" : [],
+        "model" : null
+    },
+    {
+        "name" : "Bench",
+        "description" : "-10% to job time",
+        "show" : [],
+        "building effects" : [["job time", -.20]],
+        "cost" : [],
+        "model" : null,
+    }
+]