+++ Scenes/Scripts/JobData.gd
@@ -1,0 +1,12 @@
+var job_list = [
+{
+    "name" : "Default",
+    "description": "Send your intreped heroes into the depths",
+    "requirments" : [],
+    "worker cap" : 4,
+    "exp" : 0,
+    "time" : 120,
+    "cost" : [["money", 1]],
+    "rewards" : [["minerals", 1]]
+},
+]