+++ Info/Changelog.md
@@ -1,8 +1,31 @@
 # ChangeLog
 
+## 5/25/2023
+
+buildings can now have decorations.
+
+Decorations include
+
+* Locker - job cap
+* Bin - resource
+* Bench - job time
+
+Fixes
+
+* Decorations would not appear in the tile viewer
+* Buildings would not appear in the tile viewer
+
+## 5/24/2023
+
+buildings have decorations basis now.
+
+## 5/16/2023
+
+The building data has been seperated from the tile data and is now it's own scene node.
+
 ## 5/15/2023
 
-    Tiles now can be changed form an auto loader instead of a instanced file.
+Tiles now can be changed form an auto loader instead of a instanced file.
 
 ## 5/14/2023